# docs/中医电子病历质控与标准化系统-openapi.yaml
openapi: 3.0.3
info:
  title: 中医电子病历质控与标准化系统
  description: |
    中医电子病历质控与标准化系统接口文档。

    ## 权限约定
    接口分四档权限，各接口在 description 首部以【权限：…】标注：
    - **仅管理员**：管理类操作（词典导入/回滚、数据清洗、数据集导出、日志审计、质控评分与重算）。
    - **登录即可**：只读查询与业务计算（术语查询/归一、统计看板、NLP 抽取、原始病历查看）。
    - **管理员 / 审核员**：复核相关接口（复核任务列表、提交复核）。
    - **公开**：用户登录、用户注册。
    审核员另受**数据域（行级权限）**约束：仅可读取 grade='待复核' 的病历；管理员为全库。
    执行顺序固定为「先数据域、后用户筛选」取交集，用户筛选条件不得绕过数据域。
    权限不足统一返回 HTTP 403（code=403）。
    ## 实体分层命名约定
    - **DTO**（后缀 DTO）：请求体入参模型，接口入参统一使用。
    - **VO**（后缀 VO）：响应体出参模型，接口出参统一使用。
    - **PO**（后缀 PO）：数据库持久层实体，内部使用，不暴露于接口，故本文档不列出。
    - **共享模型**（无后缀）：跨请求/响应复用的数据模型（如 StructuredData、Entity、Herb、RecordMetadata），见数据模型定义。
    ## 接口通用约定
    除标注「文件流下载」的接口外，所有接口统一返回 Result 包装 {"code":200,"msg":"string","data":any}。
    - code: 200成功；400参数非法；401未登录/token失效；403权限不足；404资源不存在；500服务异常。
    - 登录态：非登录接口需在请求头携带 Authorization: Bearer <token>。
    - 实体识别无结果统一返回空数组 []，禁止返回 null。
    - 分页统一参数：page（从1开始）、pageSize。
    - 错误响应统一结构见 ErrorVO（code/msg/data，非 200 时 data 为 null）。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地开发
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: 未登录或 token 缺失/过期（code=401）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorVO'
    Forbidden:
      description: 权限不足（code=403）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorVO'
    ServerError:
      description: 服务异常（code=500）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorVO'
  schemas:
...
    ErrorVO:
      type: object
      description: 统一错误响应（出参；非 200 时 data 为 null）
      properties:
        code:
          type: integer
          description: 错误码：400/401/403/404/500 及业务错误码
        msg:
          type: string
          description: 错误描述
        data:
          type: object
          nullable: true
          example: null
    # ========== 共享数据模型（跨请求/响应复用，非 PO） ==========
...
    LlmConfigDTO:
      type: object
      description: LLM 配置入参（PUT /api/llm/config 与 POST /api/llm/test 共用）；字段为空表示沿用当前值，apiKey 传空串或掩码表示不修改
      properties:
        enabled:
          type: boolean
          description: 是否启用 LLM；不传沿用当前
        provider:
          type: string
          enum:
            - ollama
            - openai
          description: 通道；ollama 本机免 api-key，openai 指 OpenAI 及兼容三方网关
        baseUrl:
          type: string
          description: 接口地址；留空用通道默认（ollama=http://localhost:11434，openai=https://api.openai.com/v1）
        apiKey:
          type: string
          description: api-key；空串或回传的掩码均表示不修改（明文密钥只在服务端内存中）
        model:
          type: string
          description: 模型名；留空用通道默认模型
        temperature:
          type: number
          format: double
          description: 采样温度，取值 0~2
        timeout:
          type: integer
          description: 单次请求超时（毫秒），取值 1000~300000
    LlmConfigVO:
      type: object
      description: LLM 当前生效配置（出参）；api-key 只回传掩码，明文永不出服务端
      properties:
        enabled:
          type: boolean
          description: 当前是否启用
        provider:
          type: string
          enum:
            - ollama
            - openai
        baseUrl:
          type: string
        apiKeyMask:
          type: string
          description: api-key 掩码（如 sk-****abcd）；未配置为空串
          example: sk-****abcd
        apiKeySet:
          type: boolean
          description: 是否已配置 api-key
        model:
          type: string
        temperature:
          type: number
          format: double
        timeout:
          type: integer
        available:
          type: boolean
          description: 当前是否已按生效配置装配成功（enabled 且参数合法）
    LlmTestVO:
      type: object
      description: LLM 连通性探测结果（出参）
      properties:
        provider:
          type: string
        model:
          type: string
          description: 实际使用的模型名；配置留空时为通道默认模型
        latencyMs:
          type: integer
          format: int64
          description: 往返耗时（毫秒）
        reply:
          type: string
          description: 模型回复内容（探测提示词为 ping，回复通常很短）

    # ========== 清洗与导出 DTO / VO ==========
    NormalizeDTO:
      type: object
      description: 术语标准化请求（入参）
      required:
        - type
        - term
      properties:
        type:
          type: string
          enum:
            - disease
            - symptom
            - pattern
            - herb
            - formula
          description: disease疾病 / pattern证候 / symptom症状 / herb中药 / formula方剂
        term:
          type: string
          description: 口语化/非标准术语
    NormalizeVO:
      type: object
      description: 术语标准化响应（出参；未命中时 standardTerm 返回原词、source 返回空字符串）
      properties:
        standardTerm:
          type: string
        source:
          type: string
    CleanDTO:
      type: object
      description: 数据清洗请求（入参）
      required:
        - recordIds
      properties:
        recordIds:
          type: array
          items:
            type: string
    CleanVO:
      type: object
      description: 清洗结果（出参）
      properties:
        deduped:
          type: integer
        repaired:
          type: integer
        cleared:
          type: integer
        isolated:
          type: integer
        total:
          type: integer
        normalized:
          type: integer
          description: 术语归一命中总数
        normByLevel:
          type: object
          description: 术语归一三级命中分布（精确/包含/模糊）
          properties:
            exact:
              type: integer
            contain:
              type: integer
            fuzzy:
              type: integer
    ExportDatasetDTO:
      type: object
      description: 数据集导出请求（入参）
      properties:
        format:
          type: string
          enum:
            - csv
            - json
          default: csv
        filters:
          type: object
          description: 复用基础查询条件
    # ========== 质控 DTO / VO ==========
...
    AiQueryDTO:
      type: object
      description: AI 解读/问答/复核请求（入参，共享）：interpret 用 recordId；chat 用 question（+ 可选 recordId）；review 用 recordId + correctedData（批D）
      properties:
        recordId:
          type: string
          description: 病历ID（interpret 必填；chat 可选，"这份病历…"类问题需要）
        question:
          type: string
          description: 使用者业务问题（chat 必填）
    AiReplyVO:
      type: object
      description: AI 结果（出参，共享）；规则出结论 + LLM 叙述，LLM 不可用降级且规则结论保留
      properties:
        answer:
          type: string
          description: 自然语言回答/叙述（interpret 的报告叙述 或 chat 的问答回复，恒非空）
        source:
          type: string
          description: 来源：llm（模型）/ rule（降级规则）
          enum:
            - llm
            - rule
        llmAvailable:
          type: boolean
          description: LLM 是否可用；false 时 answer 为模板/规则降级、summary 为 null
        summary:
          type: object
          nullable: true
          description: 病历要点摘要（interpret，LLM 生成）；LLM 不可用或 chat 时为 null
          properties:
            chiefComplaint:
              type: string
              description: 主诉
            diagnosis:
              type: string
              description: 诊断（中医诊断）
            syndrome:
              type: string
              description: 辨证
            prescription:
              type: string
              description: 方药
        completeness:
          type: object
          nullable: true
          description: 完整性分析（interpret，21 字段口径）；chat 时为 null
          properties:
            total:
              type: integer
              example: 21
            present:
              type: integer
            missing:
              type: array
              items:
                type: string
        coreMissing:
          type: array
          description: 核心字段缺失清单（interpret；脉象/舌象/证候/治法/方剂/中药）
          items:
            type: string
        normHits:
          type: object
          nullable: true
          description: 归一命中分布（interpret；精确/包含/模糊）；chat 时为 null
          properties:
            exact:
              type: integer
            contain:
              type: integer
            fuzzy:
              type: integer
            total:
              type: integer
        keyHints:
          type: array
          description: 关键提示（interpret；如"辨证结论为空""处方缺失"）
          items:
            type: string
        disclaimer:
          type: string
          example: AI辅助分析，最终以人工复核为准
paths:
...
  /api/governance/normalize:
    post:
      tags:
        - 清洗与导出
      summary: 术语标准化
      description: 【权限：登录即可】将口语化/非标准中医术语归一国标标准术语。未命中时 standardTerm 返回原词、source 返回空字符串
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormalizeDTO'
            example:
              type: symptom
              term: 嗓子疼
      responses:
        '200':
          description: 标准化结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/NormalizeVO'
        '400':
          description: 术语类型非法（type 不在 disease/pattern/symptom/herb/formula 范围）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/governance/stats:
    get:
      tags:
        - 清洗与导出
      summary: 治理状态统计
      description: 【权限：仅管理员】返回质控合格病历数量、待治理、已治理统计（清洗与导出页顶部状态行）
      responses:
        '200':
          description: 治理状态统计
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      qualified:
                        type: integer
                      pendingGovern:
                        type: integer
                      governedCount:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/export/dataset:
    post:
      tags:
        - 清洗与导出
      summary: 标准数据集导出
      description: 【权限：仅管理员】按条件导出质控合格的标准化病历数据集，支持 CSV/JSON 格式。文件流下载，不套 Result 包装，导出自动脱敏
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportDatasetDTO'
            example:
              format: csv
              filters: {}
      responses:
        '200':
          description: 数据集文件流（CSV为UTF-8、首行为字段名；JSON为数组格式）
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: 质控未通过，禁止导出数据集（code=2001）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/export/dataset/preview:
    post:
      tags:
        - 清洗与导出
      summary: 标准数据集预览
      description: 【权限：仅管理员】按条件返回质控合格病历的总数与前 10 条样本（供导出前预览，不下载文件）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportDatasetDTO'
      responses:
        '200':
          description: 预览结果（total + sample 前 10 条）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                      sample:
                        type: array
                        items:
                          type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/governance/clean:
    post:
      tags:
        - 清洗与导出
      summary: 数据清洗规整
      description: 【权限：仅管理员】批量清洗病历数据，完成去重、修复、规整、脏数据隔离【长耗时同步接口（3 万上限）；前端放宽超时、勿重复提交】
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CleanDTO'
      responses:
        '200':
          description: 清洗完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  msg:
                    type: string
                    example: 清洗完成
                  data:
                    $ref: '#/components/schemas/CleanVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 病历不存在（code=1006）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/ai/interpret:
    post:
      tags:
        - AI助手
      summary: AI质控解读
      description: 【权限：登录即可】对单条病历生成可解释质控解读：规则出结论（完整性/缺项重点/归一命中/关键提示），再由 LlmClient 生成叙述与要点摘要；LLM 不可用降级模板叙述。读取受数据域约束（审核员仅待复核域）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiQueryDTO'
            example:
              recordId: 3f2a1b0c-...
      responses:
        '200':
          description: 解读结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/AiReplyVO'
        '400':
          description: recordId 为空（code=400）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 病历不存在或不在数据域内（code=1006）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/ai/chat:
    post:
      tags:
        - AI助手
      summary: AI助手问答
      description: |
        【权限：登录即可】面向使用者的项目业务问答（数据/功能/流程/判定原因），不回答技术实现问题。
        路由：数据类问题→stats；归一/分布→归一命中；"这份病历…"→recordId 活跃详情；功能/流程/标准依据→知识注入；
        技术实现问题→兜底拒答。LLM 不可用→降级规则问答（source=rule）。单轮问答，无会话历史。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiQueryDTO'
            example:
              question: 当前合格率与待复核构成
              recordId: 3f2a1b0c-...
      responses:
        '200':
          description: 回答
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/AiReplyVO'
        '400':
          description: question 为空（code=400）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/ai/review:
    post:
      tags:
        - AI助手
      summary: AI复核预检意见
      description: |
        【权限：管理员 / 审核员】基于规则预检单（qc_results / 重算）与 structuredData 生成复核建议
        （疑缺脉象 / 证候-方剂冲突 / 建议复核重点）；判定仍为规则，LLM 仅叙述；
        LLM 关/挂 → 回退预检单原文。读取受数据域约束（审核员仅待复核域）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiQueryDTO'
            example:
              recordId: 3f2a1b0c-...
      responses:
        '200':
          description: 复核建议
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/AiReplyVO'
        '400':
          description: recordId 为空（code=400）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 病历不存在或不在数据域内（code=1006）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/llm/config:
    get:
      tags:
        - LLM配置
      summary: 读取 LLM 配置
      description: 【权限：仅管理员】读取当前生效的 LLM 配置。api-key 只回传掩码（如 sk-****abcd），明文密钥不出服务端
      responses:
        '200':
          description: 当前生效配置
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/LlmConfigVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - LLM配置
      summary: 覆盖 LLM 配置
      description: |
        【权限：仅管理员】覆盖运行时生效的 LLM 配置，保存即生效、无需重启。
        只作用于内存，不写回 application.yml，重启后回到配置文件基线；
        需要长期生效请写入配置文件的 llm 段。apiKey 传空串或掩码表示不修改。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmConfigDTO'
            example:
              enabled: true
              provider: ollama
              baseUrl: http://localhost:11434
              apiKey: ''
              model: qwen2.5:7b
              temperature: 0.2
              timeout: 60000
      responses:
        '200':
          description: 生效后的配置
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/LlmConfigVO'
        '400':
          description: 参数非法（通道非 ollama/openai、温度越界、超时越界，或选 openai 但未填 api-key）（code=400）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/llm/test:
    post:
      tags:
        - LLM配置
      summary: LLM 连通性探测
      description: 【权限：仅管理员】用请求体参数（省略则用当前生效配置）发一轮最小请求探测连通性，先试后存；不改变当前生效配置。探测超时上限 20 秒
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmConfigDTO'
      responses:
        '200':
          description: 连接正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/LlmTestVO'
        '400':
          description: 参数非法（通道非 ollama/openai，或选 openai 但未填 api-key）（code=400）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '502':
          description: 上游模型服务不可达或鉴权失败（code=1009，消息已脱敏可直接展示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorVO'
        '500':
          $ref: '#/components/responses/ServerError'